DB-2339: Create paginator component
