+++ starters/next-drupal-starter/pages/pages/index.js
@@ -5,8 +5,9 @@ import {
   globalDrupalStateStores,
 } from "../../lib/drupalStateContext";
 
-import Link from "next/link";
 import Layout from "../../components/layout";
+import PageHeader from "../../components/page-header.js";
+import Link from "next/link";
 
 export default function PagesList({ hrefLang, pages, multiLanguage }) {
   return (
@@ -16,9 +17,7 @@ export default function PagesList({ hrefLang, pages, multiLanguage }) {
         description="Generated by create next app."
         languageAlternates={hrefLang || false}
       />{" "}
-      <header className="prose text-2xl mx-auto mt-20">
-        <h1 className="text-center mx-auto">Pages</h1>
-      </header>
+      <PageHeader title="Pages" />
       <div className="mt-12 mx-auto max-w-[50vw]">
         <ul>
           {pages ? (
@@ -59,6 +58,9 @@ export async function getStaticProps(context) {
   });
 
   const store = getCurrentLocaleStore(locale, globalDrupalStateStores);
+
+  // clear params in case they pollute call to get node--page
+  store.params.clear();
 
   try {
     const pages = await store.getObject({
